Added a "close" protocol
Added a close protocol for when the Content-Length is not known or is 0. Adds a "X-Swiftiply-Close" header, and the server looks for <!--SC-> before trying to find the next legitmate header. Yep, it's got some problems, especially if somebody puts a <!--SC-> in their data when it's chunked. And it's probably slow having to look for that.